Move business logic into Services package
